begin refactor of file organization
