+++ lib/Brightcove/API/DI.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Brightcove\API;
+
+use Brightcove\API\API;
+use Brightcove\API\Request\IngestRequest;
+use Brightcove\API\Response\IngestResponse;
+
+class DI extends API {
+
+  protected function diRequest($method, $endpoint, $result, $is_array = FALSE, $post = NULL) {
+    return $this->client->request($method, 'ingest', $this->account, $endpoint, $result, $is_array, $post);
+  }
+
+  /**
+   * @return BrightcoveIngestResponse
+   */
+  public function createIngest($video_id, IngestRequest $request) {
+    return $this->diRequest('POST', "/videos/{$video_id}/ingest-requests", IngestResponse::class, FALSE, $request);
+  }
+}
