+++ lib/Brightcove/API/Exception/BrightcoveAuthenticationException.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Brightcove\API\Exception;
+
+class BrightcoveAuthenticationException extends Exception {}
