+++ lib/Brightcove/Object/ObjectBase.php
@@ -1,0 +1,147 @@
+<?php
+
+namespace Brightcove\Object;
+
+use Brightcove\Object\ObjectInterface;
+
+/**
+ * Base object which implements most of a needed methods to satisfy BrightcoveObject.
+ */
+class ObjectBase implements ObjectInterface {
+  /**
+   * This internal list keeps track of the changed fields.
+   *
+   * This is necessary bookeeping for patchJSON().
+   *
+   * @var array
+   */
+  private $changedFields = [];
+
+  /**
+   * Marks a field as changed.
+   *
+   * All property setters should call this function.
+   *
+   * @param string $field_name
+   */
+  protected function fieldChanged($field_name) {
+    $this->changedFields[] = $field_name;
+  }
+
+  public function postJSON() {
+    $data = [];
+    foreach ($this as $field => $val) {
+      if ($field === 'changedFields' || $val === NULL) {
+        continue;
+      }
+      if ($val instanceof BrightcoveObject) {
+        $data[$field] = $val->postJSON();
+      }
+      else if (is_array($val)) {
+        $data[$field] = [];
+        foreach ($val as $k => $v) {
+          if ($v instanceof BrightcoveObject) {
+            $data[$field][$k] = $v->postJSON();
+          }
+          else {
+            $data[$field][$k] = $v;
+          }
+        }
+      } else {
+        $data[$field] = $val;
+      }
+    }
+    return $data;
+  }
+
+  public function patchJSON() {
+    $data = [];
+    foreach ($this->changedFields as $field) {
+      $val = $this->{$field};
+      if ($val === NULL) {
+        continue;
+      }
+
+      if ($val instanceof BrightcoveObject) {
+        $data[$field] = $val->patchJSON();
+      } else if (is_array($val)) {
+        $data[$field] = [];
+        foreach ($val as $k => $v) {
+          if ($v instanceof BrightcoveObject) {
+            $data[$field][$k] = $v->patchJSON();
+          } else {
+            $data[$field][$k] = $v;
+          }
+        }
+      } else {
+        $data[$field] = $val;
+      }
+    }
+
+    $this->changedFields = [];
+
+    return $data;
+  }
+
+  public function applyJSON(array $json) {}
+
+  /**
+   * Helper method for applyJSON().
+   *
+   * Applies exactly one property on $this.
+   *
+   * @param array $json
+   *   The full JSON array, decoded as an associative array.
+   * @param string $name
+   *   The name of the property on $this.
+   * @param null|string $json_name
+   *   The name of the JSON property. If null, then it is the
+   *   same as $name.
+   * @param null|string $classname
+   *   The type of the property. If null, the JSON data will be
+   *   copied as it is. If it's a string, it will be instantiated
+   *   as a class, which implements BrightcoveObject.
+   * @param bool $is_array
+   *   If the property is an array. This will be only used
+   *   when $classname is not null.
+   */
+  protected function applyProperty(array $json, $name, $json_name = NULL, $classname = NULL, $is_array = FALSE) {
+    if ($json_name === NULL) {
+      $json_name = $name;
+    }
+    if (!isset($json[$json_name])) {
+      return;
+    }
+
+    if ($classname === NULL) {
+      $this->$name = $json[$json_name];
+    }
+    else {
+      if ($is_array) {
+        $arr = [];
+        foreach ($json[$json_name] as $k => $v) {
+          $class = new $classname();
+          $class->applyJSON($v);
+          $arr[$k] = $class;
+        }
+        $this->$name = $arr;
+      }
+      else {
+        $class = new $classname();
+        $class->applyJSON($json[$json_name]);
+        $this->$name = $class;
+      }
+    }
+  }
+
+  public static function fromJSON($json) {
+    if (is_string($json)) {
+      $json = json_decode($json, TRUE);
+    }
+
+    $ret = new static();
+    $ret->applyJSON($json);
+
+    return $ret;
+  }
+}
