+++ lib/Brightcove/Object/Video/Image.php
@@ -1,0 +1,53 @@
+<?php
+
+namespace Brightcove\Object\Video;
+
+use Brightcove\Object\ObjectBase;
+
+/**
+ * An image what represents a video, only can be thumbnail or poster.
+ */
+class Image extends ObjectBase {
+  protected $id;
+  protected $src;
+
+  public function applyJSON(array $json) {
+    parent::applyJSON($json);
+    $this->applyProperty($json, 'id');
+    $this->applyProperty($json, 'src');
+  }
+
+  /**
+   * @return string
+   */
+  public function getId() {
+    return $this->id;
+  }
+
+  /**
+   * @param string $id
+   * @return $this
+   */
+  public function setId($id) {
+    $this->id = $id;
+    $this->fieldChanged('id');
+    return $this;
+  }
+
+  /**
+   * @return string
+   */
+  public function getSrc() {
+    return $this->src;
+  }
+
+  /**
+   * @param string $src
+   * @return $this
+   */
+  public function setSrc($src) {
+    $this->src = $src;
+    $this->fieldChanged('src');
+    return $this;
+  }
+}
